% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{archive_course}
\alias{archive_course}
\title{Archive a Google Classroom Course}
\usage{
archive_course(course_id)
}
\arguments{
\item{course_id}{ID of the archived course you wish to delete}
}
\description{
Archive a Google Classroom Course
}
