% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authorize}
\alias{authorize}
\title{Authorize R package to access Google classroom API}
\usage{
authorize(token = NULL, ...)
}
\arguments{
\item{token}{an output from \code{\link{oauth2.0_token}} to set as the
authentication token.}

\item{...}{additional arguments to send to \code{\link{oauth2.0_token}}}
}
\description{
This is a function to authorize the R package to access the Googleclassroom API. If no
client.id and client.secret is provided, the package would provide predefined values.
}
\examples{
\dontrun{

authorize()
}
}
