% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{copy_form}
\alias{copy_form}
\title{Make a copy of an existing form}
\usage{
copy_form(form_id)
}
\arguments{
\item{form_id}{The form_id that is desired to be copied.}

\item{description}{The description for the new form as a string.}

\item{document_title}{The title for the form file that will be stored in Google Drive}
}
\description{
Make a copy of an existing form
}
\examples{
\dontrun{
#'
# Make the form
form_info <- create_form(form_id = <google_form_url>)
}
}
