% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{create_course}
\alias{create_course}
\title{Create a new course}
\usage{
create_course(
  owner_id = get_owner_id()$id,
  name = "New course",
  load_url = TRUE
)
}
\arguments{
\item{owner_id}{The ownerId to use to create the course. Will attempt to retrieve ownerId based on credentials with get_owner_id()}

\item{name}{Name of the new course. Required.}

\item{load_url}{Load URL into an HTML Browser}
}
\value{
List of metadata of newly created course.
}
\description{
Create a new course
}
\examples{
\dontrun{

owner_id <- get_owner_id()
course_df <- create_course(name = "Computer Science 101")
}
}
