% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{create_coursework}
\alias{create_coursework}
\title{Create a new coursework}
\usage{
create_coursework(
  course_id = NULL,
  topic_id = NULL,
  name = NULL,
  work_type = NULL,
  due_date = NULL,
  description = NULL,
  full_response = FALSE
)
}
\arguments{
\item{course_id}{Course id of where to make the new coursework. Can find from end of URL e.g. "https://classroom.google.com/c/COURSE_ID_IS_HERE"}

\item{name}{Name of new coursework}

\item{full_response}{Parameter to decide whether to return the full response or just the presentation ID}
}
\description{
Create a new coursework
}
