% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{create_form}
\alias{create_form}
\title{Create a new form}
\usage{
create_form(
  title = NULL,
  description = "",
  is_quiz = FALSE,
  full_response = TRUE
)
}
\arguments{
\item{full_response}{Parameter to decide whether to return the full response or just the presentation ID}

\item{form_id}{Forms id of where to make the new coursework. Can find from end of URL e.g. "https://classroom.google.com/c/COURSE_ID_IS_HERE"}

\item{name}{Name of new coursework}
}
\description{
Create a new form
}
