% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{create_topic}
\alias{create_topic}
\title{Create a new topic}
\usage{
create_topic(course_id = NULL, name = NULL, full_response = FALSE)
}
\arguments{
\item{course_id}{Course id of where to make the new topic. Can find from end of URL e.g. "https://classroom.google.com/c/COURSE_ID_IS_HERE"}

\item{name}{Name of new topic}

\item{full_response}{Parameter to decide whether to return the full response or just the presentation ID}
}
\description{
Create a new topic
}
