% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_handler}
\alias{date_handler}
\title{Handle and parse a due_date}
\usage{
date_handler(due_date)
}
\arguments{
\item{due_date}{A string that is a date in format of year-month-day}
}
\description{
Handle and parse a due_date
}
\examples{
\dontrun{
date_handler("2025-12-1")
}
}
