% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{delete_course}
\alias{delete_course}
\title{Delete a Google Classroom Course}
\usage{
delete_course(course_id)
}
\arguments{
\item{course_id}{ID of the archived course you wish to delete}
}
\description{
Delete a Google Classroom Course
}
