% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{get_course_properties}
\alias{get_course_properties}
\title{Get Google Classroom Course Properties}
\usage{
get_course_properties(course_id)
}
\arguments{
\item{course_id}{ID of the course you wish to retrieve information from}
}
\description{
Get Google Classroom Course Properties
}
