% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{get_coursework_list}
\alias{get_coursework_list}
\title{Get list of courseworks for a course}
\usage{
get_coursework_list(course_id)
}
\arguments{
\item{course_id}{ID of the course to retrieve the courseworks from}
}
\description{
Get list of courseworks for a course
}
\examples{
\dontrun{

course_id <- get_course_list()$courses$id[1]

get_coursework_list(course_id)
}
}
