% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{get_coursework_properties}
\alias{get_coursework_properties}
\title{Get Google Classroom Course Properties}
\usage{
get_coursework_properties(course_id, coursework_id)
}
\arguments{
\item{course_id}{ID of the course you wish to retrieve information about a particular coursework}

\item{coursework_id}{ID of the coursework you wish to retrieve information about}
}
\description{
Get Google Classroom Course Properties
}
