% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{get_form_properties}
\alias{get_form_properties}
\title{Get Google Classroom Course Properties}
\usage{
get_form_properties(form_id)
}
\arguments{
\item{id}{ID of the course}
}
\description{
Get Google Classroom Course Properties
}
