% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-responses.R
\name{get_form_responses}
\alias{get_form_responses}
\title{Get form responses}
\usage{
get_form_responses(form_id = NULL, form_url = NULL)
}
\arguments{
\item{form_id}{The id of the google form to be updated}

\item{form_url}{Google form url}
}
\description{
Get form responses
}
\examples{
\dontrun{

form_url <- "https://docs.google.com/forms/d/1woWtLVviIhrwb-NYEjVMO_IV2-62vOhaS4N0/edit#responses"
get_form_responses(form_url = form_url)
}
}
