% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{get_linked_quizzes_list}
\alias{get_linked_quizzes_list}
\title{Get list of forms used in a course as quizzes}
\usage{
get_linked_quizzes_list(course_id)
}
\arguments{
\item{course_id}{ID of the course to retrieve the linked quizzes from}
}
\description{
Get list of forms used in a course as quizzes
}
\examples{
\dontrun{
course_id <- get_course_list()$courses$id[1]
quiz_list <- get_linked_quizzes_list(course_id)
}
}
