% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{get_materials_list}
\alias{get_materials_list}
\title{Get list of materials for a course}
\usage{
get_materials_list(course_id)
}
\arguments{
\item{id}{ID of the course}
}
\description{
Get list of materials for a course
}
\examples{
\dontrun {
course_id <- get_course_list()$courses$id[3]
materials_df <- get_materials_list(course_id)
}
}
