% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{get_materials_properties}
\alias{get_materials_properties}
\title{Get Google Classroom Materials properties}
\usage{
get_materials_properties(course_id, materials_id)
}
\arguments{
\item{materials_id}{The material id you wish to retrieve}

\item{id}{ID of the course}
}
\description{
Get Google Classroom Materials properties
}
