% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{get_topic_list}
\alias{get_topic_list}
\title{Get list of topics for a course}
\usage{
get_topic_list(course_id)
}
\arguments{
\item{id}{ID of the course}
}
\description{
Get list of topics for a course
}
\examples{
\dontrun{

course_id <- get_course_list()$courses$id[1]

get_topic_list(course_id)
}
}
