% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{get_topic_properties}
\alias{get_topic_properties}
\title{Get Google Classroom Topic Properties}
\usage{
get_topic_properties(course_id, topic_id)
}
\arguments{
\item{course_id}{ID of the course}

\item{topic_id}{ID of the course}
}
\description{
Get Google Classroom Topic Properties
}
