% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms-request.R
\name{is.google_forms_request}
\alias{is.google_forms_request}
\title{Check if the object is a google forms request object}
\usage{
is.google_forms_request(x)
}
\arguments{
\item{x}{A google_forms_request object created from the rgoogleclassroom package}
}
\description{
Check if the object is a google forms request object
}
