% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{publish_coursework}
\alias{publish_coursework}
\title{Publish a Google Classroom CourseWork}
\usage{
publish_coursework(course_id, coursework_id)
}
\arguments{
\item{course_id}{ID of the archived course you wish to delete}

\item{coursework_id}{coursework ID of the coursework you wish to publish}
}
\description{
Publish a Google Classroom CourseWork
}
