% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{time_handler}
\alias{time_handler}
\title{Handle and parse a time}
\usage{
time_handler(due_time = lubridate::hms("23:59:59"))
}
\arguments{
\item{due_time}{A string that is a date in format of year-month-day. Default is midnight.}
}
\description{
Handle and parse a time
}
\examples{
\dontrun{
time_handler("21:30:59")
}
}
