% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottrpal-translate.R
\name{translate_questions_api}
\alias{translate_questions_api}
\title{Translate Markua questions for submission to Google API}
\usage{
translate_questions_api(quiz_path)
}
\arguments{
\item{quiz_path}{The file path to a Markua formatted quiz to be translated to Google Forms API format}

\item{verbose}{Whether progress messages should be given.}
}
\value{
A list of the output from [ottrpal::check_question] with messages/warnings regarding each question and each check.
#'
}
\description{
Takes a Markua formatted quiz and translates it to something that can be sent to Google Forms API.
}
\examples{
\dontrun{

# Using quiz example

quiz_path <- markdown_quiz_path()
good_quiz <- readLines(quiz_path)
good_quiz_specs <- parse_quiz(good_quiz)
good_quiz_checks <- check_all_questions(good_quiz_specs)
}

}
