% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-edits.R
\name{update_form_settings}
\alias{update_form_settings}
\title{Create a multiple choice question}
\usage{
update_form_settings(
  form_id = NULL,
  title = NULL,
  description = NULL,
  google_forms_request = NULL,
  quiet = FALSE
)
}
\arguments{
\item{form_id}{The id of the google form to be updated}

\item{title}{An updated title}

\item{description}{An updated description}

\item{google_forms_request}{A google forms request object. If not supplied, it will be created new.}

\item{quiet}{TRUE/FALSE you'd like a progress message?}
}
\description{
Create a multiple choice question
}
\examples{
\dontrun{

update_form_settings(
  form_id = "12345",
  new_title = NULL,
  new_description = NULL
)
}
}
