% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchUpdate.R
\name{commit_to_form}
\alias{commit_to_form}
\title{Commit changes to a Google form}
\usage{
commit_to_form(form_id, google_forms_request)
}
\arguments{
\item{form_id}{The id of the google form to be updated}

\item{google_forms_request}{The google slide request to be applied to the slides}
}
\description{
Commit changes to a Google form
}
