% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{create_course}
\alias{create_course}
\title{Create a new course}
\usage{
create_course(title = NULL, full_response = FALSE)
}
\arguments{
\item{title}{Title of the course}

\item{full_response}{Parameter to decide whether to return the full response or just the presentation ID}
}
\description{
Create a new course
}
