% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{create_form}
\alias{create_form}
\title{Create a new form}
\usage{
create_form(title = NULL, document_title = "new_form", description = "")
}
\arguments{
\item{title}{The title for the new form. Required as a string.}

\item{document_title}{The title for the form file that will be stored in Google Drive}

\item{description}{The description for the new form as a string.}
}
\description{
Create a new form
}
\examples{
\dontrun{
#'
# Make the form
form_info <- create_form(
  title = "A great quiz",
  description = "This quiz is tricky"
)
}
}
