% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{create_material}
\alias{create_material}
\title{Create a new material}
\usage{
create_material(
  course_id = NULL,
  topic_id = NULL,
  title = NULL,
  due_date = NULL,
  description = NULL,
  material_link = NULL,
  full_response = FALSE
)
}
\arguments{
\item{course_id}{Course id of where to make the new materials. Can find from end of URL e.g. "https://classroom.google.com/c/COURSE_ID_IS_HERE"}

\item{title}{Name of new material}

\item{description}{A description for the new material}

\item{material_link}{A URL to go with the associated material}

\item{full_response}{Parameter to decide whether to return the full response or just the presentation ID}
}
\description{
Create a new material
}
\examples{
\dontrun {
course_id <- get_course_list()$courses$id[3]
topic_id <- get_topic_list(course_id)$topic$topicId[1]

create_material(course_id, topic_id, title = "new material")

}
}
