% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-edits.R
\name{create_multiple_choice_question}
\alias{create_multiple_choice_question}
\title{Commit changes to a Google form}
\usage{
create_multiple_choice_question(
  question_kind = "choiceQuestion",
  required = FALSE,
  question = NULL,
  choice_vector = NULL,
  shuffle_opt = FALSE,
  correct_answer = NULL,
  point_value = NULL
)
}
\arguments{
\item{question_kind}{Can be choiceQuestion or textQuestion}

\item{required}{TRUE or FALSE is this a required question? Default is not required.}

\item{question}{a string that is what the question should say}

\item{choice_vector}{a character vector of the choices that should be given for this question}

\item{shuffle_opt}{TRUE or FALSE options should be shuffled? default is FALSE}

\item{correct_answer}{The index that corresponds to the correct answer in the `choice_vector` supplied}

\item{form_id}{The id of the google form to be updated}
}
\description{
Commit changes to a Google form
}
\examples{
\dontrun{

create_multiple_choice_question(
  question = "What answer do you want?", 
  choice_vector = c("A", "B", "C", "D")
  )

}
}
