% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form-edits.R
\name{create_text_question}
\alias{create_text_question}
\title{Create a text question}
\usage{
create_text_question(
  form_id = NULL,
  question_kind = "textQuestion",
  required = FALSE,
  question = NULL,
  correct_answer = NULL,
  google_forms_request = NULL,
  point_value = 1,
  location = 0
)
}
\arguments{
\item{form_id}{The id of the google form to be updated}

\item{required}{TRUE or FALSE is this a required question? Default is not required.}

\item{question}{a string that is what the question should say}

\item{correct_answer}{a vector that has answers that would be considered correct}

\item{google_forms_request}{A google forms request object. If not supplied, it will be created new.}

\item{point_value}{An integer representing how many points}

\item{location}{Where should the new question be added}

\item{choice_vector}{a character vector of the choices that should be given for this question}

\item{shuffle_opt}{TRUE or FALSE options should be shuffled? default is FALSE}
}
\description{
This function makes a google request object that will be able to be posted with a batch request and
and added to a Google form to edit it.
}
\examples{
\dontrun{

create_text_question(
  form_id = "12345",
  question = "Put text here",
  correct_answer = c("dog", "cat", "fish", "elephant"),
  shuffle_opt = TRUE
)
}
}
