% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursework.R
\name{delete_coursework}
\alias{delete_coursework}
\title{Delete a Google Classroom Coursework}
\usage{
delete_coursework(course_id, coursework_id)
}
\arguments{
\item{coursework_id}{ID of the archived course you wish to delete}
}
\description{
Delete a Google Classroom Coursework
}
