% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive.R
\name{get_drive_file_list}
\alias{get_drive_file_list}
\title{Get list of files from a Google Shared Drive}
\usage{
get_drive_file_list(drive_id)
}
\arguments{
\item{id}{ID of the course}
}
\description{
Get list of files from a Google Shared Drive
}
