% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{get_form_properties}
\alias{get_form_properties}
\title{Get Google Form Properties}
\usage{
get_form_properties(form_id)
}
\arguments{
\item{form_id}{Google form Id}
}
\description{
Get Google Form Properties
}
\examples{
\dontrun{

# Make the form
form_info <- create_form(title = "A great quiz", description = "This quiz is tricky")

# Get info about the form
form_info <- get_form_properties(form_id = form_info$formId)
}
}
