% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{get_materials_properties}
\alias{get_materials_properties}
\title{Get Google Classroom Materials properties}
\usage{
get_materials_properties(course_id, materials_id)
}
\arguments{
\item{course_id}{ID of the course}

\item{materials_id}{The material id you wish to retrieve}
}
\description{
Get Google Classroom Materials properties
}
\examples{
\dontrun{
course_id <- get_course_list()$courses$id[3]
materials_id <- get_materials_list(course_id)$material_id$courseWorkMaterial$id[1]

get_materials_properties(course_id, materials_id)

}
}
