% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{make_form_quiz}
\alias{make_form_quiz}
\title{Turn a form into a quiz}
\usage{
make_form_quiz(form_id)
}
\arguments{
\item{form_id}{The id of the google form to be updated into a Quiz}
}
\description{
Turn a form into a quiz
}
\examples{
\dontrun{

# Make the form
form_info <- create_form(title = quiz_title)

# Now make it a quiz
make_form_quiz(form_id = form_info$formId)
}
}
