% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottrpal-translate.R
\name{ottr_quiz_to_google}
\alias{ottr_quiz_to_google}
\title{Create quiz batch request from Markua quiz}
\usage{
ottr_quiz_to_google(course_id = NULL, form_id = NULL, make_new_quiz = FALSE)
}
\arguments{
\item{form_id}{quiz_specs which is output from [ottrpal::parse_quiz].}

\item{quiz_name}{The name of the quiz being checked.}

\item{verbose}{Whether progress messages should be given.}

\item{ignore_coursera}{Coursera doesn't like `!` or `:` in the quizzes. Do not convert quizzes to coursera and ignore ! and : in question prompts that would not be allowed in Leanpub quizzes when converted to a Coursera quiz. Default is to ignore Coursera compatibility.}
}
\value{
A list of the output from [ottrpal::check_question] with messages/warnings regarding each question and each check.
#'
}
\description{
Takes output from [ottrpal::parse_quiz] and runs checks on each question in a quiz by calling [ottrpal::check_question] for each question.
First splits questions into their own data frame. Returns a list of messages/warnings about each question's set up.
}
\examples{
\dontrun{

# Using quiz example

quiz_path <- markdown_quiz_path()

ottr_quiz_to_google(quiz_path)
}

}
