% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottrpal-translate.R
\name{ottr_quiz_to_google}
\alias{ottr_quiz_to_google}
\title{Create quiz batch request from Markua quiz}
\usage{
ottr_quiz_to_google(
  quiz_path = NULL,
  course_id = NULL,
  form_id = NULL,
  make_new_quiz = FALSE
)
}
\arguments{
\item{quiz_path}{file path to a markdown Markua quiz}

\item{course_id}{An id for the course where this is to be published and linked.}

\item{form_id}{form id where this quiz is to be published. Alternatively, if you want a new quiz to be made, you should set make_new_quiz = TRUE and leave this NULL.}

\item{make_new_quiz}{This can only be used if form_id is not specified. This will make a new quiz}
}
\description{
Takes a file path to a Markua formatted quiz and
}
\examples{
\dontrun{

# Using quiz example

quiz_path <- markdown_quiz_path()

ottr_quiz_to_google(quiz_path)
}

}
