% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottrpal-translate.R
\name{translate_questions_api}
\alias{translate_questions_api}
\title{Translate Markua questions for submission to Google API}
\usage{
translate_questions_api(quiz_path, output_path = NULL)
}
\arguments{
\item{quiz_path}{The file path to a Markua formatted quiz to be translated to Google Forms API format}

\item{output_path}{Optional file path to save the formatted data to a JSON file}
}
\value{
A list of the output from [ottrpal::check_question] with messages/warnings regarding each question and each check.
#'
}
\description{
Takes a Markua formatted quiz and translates it to something that can be sent to Google Forms API.
}
\examples{
\dontrun{

# Using quiz example

quiz_path <- markdown_quiz_path()
parsed_questions <- translate_questions_api(quiz_path)
}

}
